+++ Cargo.toml
@@ -8,13 +8,12 @@ build = "build.rs"
 lazy_static = "0.2.1"
 regex = "0.1.71"
 rustyline = "1.0.0"
-easy_ll = { path = "easy_ll", version = "^0.1.0" }
-weld_common = { path = "weld_common", version = "^0.1.0" }
 libc = "0.2.0"
 csv = "0.15.0"
 log = "*"
 env_logger = "*"
 chrono = "*"
+llvm-sys = "38.0.1"
 
 [lib]
 path = "weld/lib.rs"
