+++ build.rs
@@ -1,3 +1,4 @@
+use std::env;
 use std::process::Command;
 
 fn main() {
@@ -28,5 +29,12 @@ fn main() {
         .status()
         .unwrap();
 
+    let target = env::var("TARGET").unwrap();
+    if target == "x86_64-apple-darwin" {
+        let libs = vec!["z", "c++"];
+        for lib in libs {
+            println!("cargo:rustc-link-lib={}", lib);
+        }
+    }
     println!("cargo:rustc-link-lib=dylib=stdc++");
 }
