--- easy_ll/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "easy_ll"
-version = "0.1.0"
-authors = ["Matei Zaharia <matei.zaharia@gmail.com>"]
-build = "build.rs"
-
-[dependencies]
-llvm-sys = "38.0.1"
-libc = "0.2.0"
-log = "*"
