--- easy_ll/build.rs
@@ -1,11 +1,0 @@
-use std::env;
-
-fn main() {
-    let target = env::var("TARGET").unwrap();
-    if target == "x86_64-apple-darwin" {
-        let libs = vec!["z", "c++"];
-        for lib in libs {
-            println!("cargo:rustc-link-lib={}", lib);
-        }
-    }
-}
