--- weld_common/Cargo.toml
@@ -1,4 +1,0 @@
-[package]
-name = "weld_common"
-version = "0.1.0"
-authors = ["Shoumik Palkar <shoumikpalkar@gmail.com>"]
